# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ControlWeave API
  description: GRC platform API for managing compliance frameworks, controls, evidence, and audit logs.
  version: 2.4.3
  contact:
    name: ControlWeave Support
    url: https://controlweave.com
servers:
  - url: /
    description: Default server
paths:
...
  /api/v1/risks/{id}/evidence:
    post:
      summary: Attach evidence to a risk
      description: >
        relevance records why this document is evidence for this risk, because
        the same document supports different risks for different reasons. An
        unrecognized value returns 400 naming the valid options rather than a
        database constraint error.
      operationId: linkRiskEvidence
      tags: [Risk Register]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [evidenceId]
              properties:
                evidenceId: { type: string, format: uuid }
                relevance:
                  type: string
                  enum: [assessment, treatment, monitoring, acceptance]
                  default: monitoring
                notes: { type: string }
      responses:
        '201': { description: Attached }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/risks/{id}/evidence/{evidenceId}:
    delete:
      summary: Detach evidence from a risk
      operationId: unlinkRiskEvidence
      tags: [Risk Register]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: evidenceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Detached }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: array
          items:
            type: string
        correlationId:
          type: string
...